+++ src/main/java/com/jlzb/storedownload/http/OppoHttpRequest.java
@@ -51,8 +51,10 @@ public class OppoHttpRequest extends HttpRequest {
     @Override
     public void download(AppInfo appInfo) {
         HttpClient httpClient = null;
+        FileOutputStream fileOutputStream = null;
+        File f = null;
         try {
-            File f = new File(appInfo.toString() + ".temp.apk");
+            f = new File(appInfo.toString() + ".temp.apk");
             if (!f.exists())
                 f.createNewFile();
 
@@ -60,18 +62,38 @@ public class OppoHttpRequest extends HttpRequest {
             HttpGet httpGet = createHttpGet();
             HttpResponse response = httpClient.execute(httpGet);
 
-            response.getEntity().writeTo(new FileOutputStream(f));
+            fileOutputStream = new FileOutputStream(f);
+            response.getEntity().writeTo(fileOutputStream);
 
-            //删除临时下载文件
-            f.delete();
+            /*inputStream = response.getEntity().getContent();
+            fileOutputStream = new FileOutputStream(f);
+
+            byte[] bytes = new byte[1024];
+            int lenght = 0;
+
+            while((lenght = inputStream.read()) != -1) {
+                fileOutputStream.write(bytes, 0 , lenght);
+            }*/
 
             //每次下载完休息2秒
             //Thread.sleep(RuleManage.rule.getSleeptime());
         } catch (Exception e) {
             e.printStackTrace();
         } finally {
+            if(fileOutputStream != null)
+                try {
+                    fileOutputStream.flush();
+                    fileOutputStream.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
             if(httpClient != null)
                 httpClient.getConnectionManager().shutdown();
+
+            //删除临时下载文件
+            if(f != null && f.exists())
+                f.delete();
         }
     }
